Feat/cart flow (#5)
* add new country, languages

* New cart flow

+++ src/Manager/SessionManager.php
@@ -44,10 +44,11 @@ class SessionManager
 
     public function removeSession(UserInterface $user, Session $session)
     {
-        $session = $this->entityManager->getRepository(Session::class)
-            ->findSessionByCustomer($session, $user);
-
-        $this->entityManager->remove($session);
-        $this->entityManager->flush();
+        if ($session->getCustomers()->contains($user)) {
+            $session = $this->entityManager->getRepository(Session::class)
+                ->findSessionByCustomer($session, $user);
+            $this->entityManager->remove($session);
+            $this->entityManager->flush();
+        }
     }
 }
